requirements.txt
asgiref==3.11.0
Django==6.0
djangorestframework==3.16.1
dotenv==0.9.9
greenlet==3.3.0
playwright==1.57.0
psycopg==3.3.2
psycopg-binary==3.3.2
pyee==13.0.0
python-dotenv==1.2.1
sqlparse==0.5.5
typing_extensions==4.15.0
